--- src/components/common/NavigationDrawer/NavigationDrawer.module.css
@@ -1,111 +1,0 @@
-.container {
-  width: 360px;
-  padding: 12px;
-  background-color: var(--c-surface);
-  border-radius: 0 16px 16px 0;
-}
-
-.labelText {
-  font-size: var(--t-label-large-size);
-  font-weight: var(--t-label-large-weight);
-  line-height: var(--t-label-large-line-height);
-  color: inherit;
-  letter-spacing: var(--t-label-large-letter-spacing);
-}
-
-.indicator {
-  position: relative;
-  display: flex;
-  gap: 12px;
-  align-items: center;
-  width: 336px;
-  height: 56px;
-  padding: 0 24px 0 16px;
-  color: rgb(var(--c-on-surface-variant-rgb));
-  background-color: var(--c-surface);
-  border-radius: 28px;
-}
-
-.indicator:focus-visible {
-  outline: 0;
-}
-
-.indicatorLayer {
-  position: absolute;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  border-radius: 28px;
-}
-
-.indicator:hover > .indicatorLayer {
-  background-color: rgb(var(--c-on-surface-variant-rgb) / 16%);
-}
-
-.indicator:focus > .indicatorLayer {
-  background-color: rgb(var(--c-on-surface-variant-rgb) / 24%);
-}
-
-.indicator:active > .indicatorLayer {
-  background-color: rgb(var(--c-on-surface-variant-rgb) / 20%);
-}
-
-.indicator[aria-selected="true"] {
-  color: rgb(var(--c-on-secondary-container-rgb));
-  background-color: var(--c-secondary-container);
-}
-
-.indicator[aria-selected="true"]:hover > .indicatorLayer {
-  background-color: rgb(var(--c-on-secondary-container-rgb) / 16%);
-}
-
-.indicator[aria-selected="true"]:focus > .indicatorLayer {
-  background-color: rgb(var(--c-on-secondary-container-rgb) / 24%);
-}
-
-.indicator[aria-selected="true"]:active > .indicatorLayer {
-  background-color: rgb(var(--c-on-secondary-container-rgb) / 20%);
-}
-
-.spacer {
-  flex: 1 1 0;
-}
-
-.badgeLabelText {
-  font-size: var(--t-label-large-size);
-  font-weight: var(--t-label-large-weight);
-  line-height: var(--t-label-large-line-height);
-  color: rgb(var(--c-on-surface-variant-rgb));
-  letter-spacing: var(--t-label-large-letter-spacing);
-}
-
-@media screen and (min-width: 0) and (max-width: 904px) {
-  .container {
-    position: absolute;
-    top: 0;
-    left: 0;
-    z-index: 2;
-    min-height: 100vh;
-    box-shadow: var(--e-level1);
-  }
-
-  .scrim {
-    position: absolute;
-    top: 0;
-    z-index: 1;
-    width: 100%;
-    height: 100%;
-    background-color: rgb(50 47 55 / 40%);
-  }
-}
-
-@media screen and (min-width: 905px) {
-  .container {
-    min-height: calc(100vh - 64px);
-  }
-
-  .scrim {
-    display: none;
-  }
-}
