--- src/components/common/NavigationDrawer/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./NavigationDrawer";
