--- src/components/common/TopAppBar/TopAppBar.module.css
@@ -1,106 +1,0 @@
-.container {
-  display: flex;
-  gap: 4px;
-  align-items: center;
-  height: 64px;
-  padding: 0 4px;
-  background-color: var(--c-surface);
-}
-
-.navigationIcon {
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 48px;
-  height: 48px;
-  color: rgb(var(--c-on-surface-rgb));
-  background-color: var(--c-surface);
-  border-radius: 14px;
-}
-
-.navigationIconLayer {
-  position: absolute;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  border-radius: 14px;
-}
-
-.navigationIcon:hover > .navigationIconLayer {
-  background-color: rgb(
-    var(--c-on-surface-rgb) / var(--s-hover-state-layer-opacity)
-  );
-}
-
-.navigationIcon:focus > .navigationIconLayer {
-  background-color: rgb(
-    var(--c-on-surface-rgb) / var(--s-focus-state-layer-opacity)
-  );
-}
-
-.navigationIcon:active > .navigationIconLayer {
-  background-color: rgb(
-    var(--c-on-surface-rgb) / var(--s-pressed-state-layer-opacity)
-  );
-}
-
-.headline {
-  font-size: var(--t-title-large-size);
-  font-weight: var(--t-title-large-weight);
-  line-height: var(--t-title-large-line-height);
-  color: rgb(var(--c-on-surface-rgb));
-  letter-spacing: var(--t-title-large-letter-spacing);
-}
-
-.ml12 {
-  margin-left: 12px;
-}
-
-.spacer {
-  flex: 1 1 0;
-}
-
-.trailingIcons {
-  display: flex;
-}
-
-.interactiveIcon {
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 48px;
-  height: 48px;
-  color: rgb(var(--c-on-surface-variant-rgb));
-  background-color: transparent;
-  border-radius: 14px;
-}
-
-.interactiveIconLayer {
-  position: absolute;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  border-radius: 14px;
-}
-
-.interactiveIcon:hover > .interactiveIconLayer {
-  background-color: rgb(
-    var(--c-on-surface-variant-rgb) / var(--s-hover-state-layer-opacity)
-  );
-}
-
-.interactiveIcon:focus > .interactiveIconLayer {
-  background-color: rgb(
-    var(--c-on-surface-variant-rgb) / var(--s-focus-state-layer-opacity)
-  );
-}
-
-.interactiveIcon:active > .interactiveIconLayer {
-  background-color: rgb(
-    var(--c-on-surface-variant-rgb) / var(--s-pressed-state-layer-opacity)
-  );
-}
