+++ src/components/features/TopAppBar/TopAppBar.module.css
@@ -1,0 +1,20 @@
+.container {
+    display: flex;
+    gap: 4px;
+    align-items: center;
+    height: 64px;
+    padding: 0 4px;
+    background-color: var(--c-surface);
+}
+
+.headline {
+    font-size: var(--t-title-large-size);
+    font-weight: var(--t-title-large-weight);
+    line-height: var(--t-title-large-line-height);
+    color: rgb(var(--c-on-surface-rgb));
+    letter-spacing: var(--t-title-large-letter-spacing);
+}
+
+.spacer {
+    flex: 1 1 0;
+}
